Rename cryptoOperations.js to station-id.js
